requirements.txt
argcomplete==3.1.4
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contextlib2==21.6.0
docopt==0.6.2
Flask==3.0.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==7.0.0
internetarchive==3.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
numpy==1.26.1
opencv-python==4.8.1.78
openlibrary-client @ git+https://github.com/internetarchive/openlibrary-client.git@f15ee489ad2718f4276fad58bca78b5844f5b32d
packaging==23.2
pipx==1.2.1
psycopg2==2.9.9
pyzbar==0.1.9
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
schema==0.7.5
simplejson==3.19.2
tqdm==4.66.1
urllib3==2.0.7
userpath==1.9.1
Werkzeug==3.0.1
zipp==3.17.0
